+++ OpenCOOD/opencood/models/sub_modules/base_bev_backbone.py
@@ -102,7 +102,7 @@ class BaseBEVBackbone(nn.Module):
                         )
                     )
                 else:
-                    stride = np.round(1 / stride).astype(np.int)
+                    stride = int(np.round(1 / stride))
                     self.deblocks.append(
                         nn.Sequential(
                             nn.Conv2d(num_filters[idx], num_upsample_filters[idx], stride, stride=stride, bias=False),
